pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Hello World"
license = "GNU Affero"
authors = ["Nicholas M. Synovic <nicholas.synovic@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/NicholasSynovic/template_python"
repository = "https://github.com/NicholasSynovic/template_python"
keywords = ["Template"]
classifiers = ["Development Status :: 1 - Planning", "Programming Language :: Python :: 3.9"]

[tool.poetry.dependencies]
python = "^3.10"
modin = {version = "^0.32.0", extras = ["all"]}
sqlalchemy = "^2.0.35"
click = "^8.1.7"
openvino = "^2024.4.0"
sentence-transformers = "^3.1.1"
torch = {version = "^2.4.1+cu124", source = "pytorch"}
torchvision = {version = "^0.19.1+cu124", source = "pytorch"}
torchaudio = {version = "^2.4.1+cu124", source = "pytorch"}
optimum = {version = "^1.22.0", extras = ["openvino", "nncf"]}
langchain = "^0.3.1"
langchain-huggingface = "^0.1.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
